+++ internal/image/image_data_test.go
@@ -8,18 +8,30 @@ import (
 
 var (
 	images = []string{
-		"test_data/image_cast0.tif",      // Int16[-10000, 10000] -> [-1,1]
-		"test_data/image_cast1.tif",      // Float32[-1, 1] -> [-1,1]
-		"test_data/image_cast2.tif",      // UInt8[-127, 127] -> [-1,1]
-		"test_data/image_cast3.tif",      // UInt8[-127, 127] -> [0,0.5]
-		"test_data/image_cast4.tif",      // UInt8[-127, 127] ->^2 [-1,1]
-		"test_data/image_cast5.tif",      // UInt8[-127, 127] ->^2 [0,0.5]
-		"test_data/image_cast6.tif",      // UInt8[-127, 127] -> [0,1] (no value [0.5,1])
-		"test_data/image_cast7.tif",      // Int16[-10000, 10000] ->^2 [0,1]
-		"test_data/image_20180812_1.tif", // Float32[0, 1] -> [0,1]
-		"test_data/image_20180812_2.tif", // Float32[0, 1] -> [0,1]
-		"test_data/image_20180812_3.tif", // Float32[0, 1] -> [0,1]
-		"test_data/image_20180812.tif",   // Float32[0, 1] -> [0,1]
+		"test_data/image_cast0.tif",  // Int16[-10000, 10000] -> [-1,1]
+		"test_data/image_cast1.tif",  // Float32[-1, 1] -> [-1,1]
+		"test_data/image_cast2.tif",  // UInt8[-127, 127] -> [-1,1]
+		"test_data/image_cast3.tif",  // UInt8[-127, 127] -> [0,0.5]
+		"test_data/image_cast4.tif",  // UInt8[-127, 127] ->^2 [-1,1]
+		"test_data/image_cast5.tif",  // UInt8[-127, 127] ->^2 [0,0.5]
+		"test_data/image_cast6.tif",  // UInt8[-127, 127] -> [0,1] (no value [0.5,1])
+		"test_data/image_cast7.tif",  // Int16[-10000, 10000] ->^2 [0,1]
+		"test_data/image_warp0.tif",  // Float32[0, 1] -> [0,1]
+		"test_data/image_warp1.tif",  // Float32[0, 1] -> [0,1]
+		"test_data/image_warp2.tif",  // Float32[0, 1] -> [0,1] (nodata=-1)
+		"test_data/image_warp3.tif",  // Float32[0, 1] -> [0,1]
+		"test_data/image_warp43.vrt", // Float32[0, 1] -> [0,1] (2 bands: an_image+image_warp3)
+		"test_data/image_warp01.vrt", // Float32[0, 1] -> [0,1] (2 bands: an_image+image_warp1)
+	}
+
+	imageDFormatFloat32NoData0 = geocube.DataMapping{
+		DataFormat: geocube.DataFormat{
+			DType:  geocube.DTypeFLOAT32,
+			NoData: 0,
+			Range:  geocube.Range{Min: 0, Max: 1},
+		},
+		RangeExt: geocube.Range{Min: 0, Max: 1},
+		Exponent: 1,
 	}
 	imagesDFormat = []geocube.DataMapping{
 		{
@@ -94,24 +106,8 @@ var (
 			RangeExt: geocube.Range{Min: 0, Max: 1},
 			Exponent: 2,
 		},
-		{
-			DataFormat: geocube.DataFormat{
-				DType:  geocube.DTypeFLOAT32,
-				NoData: 0,
-				Range:  geocube.Range{Min: 0, Max: 1},
-			},
-			RangeExt: geocube.Range{Min: 0, Max: 1},
-			Exponent: 1,
-		},
-		{
-			DataFormat: geocube.DataFormat{
-				DType:  geocube.DTypeFLOAT32,
-				NoData: 0,
-				Range:  geocube.Range{Min: 0, Max: 1},
-			},
-			RangeExt: geocube.Range{Min: 0, Max: 1},
-			Exponent: 1,
-		},
+		imageDFormatFloat32NoData0,
+		imageDFormatFloat32NoData0,
 		{
 			DataFormat: geocube.DataFormat{
 				DType:  geocube.DTypeFLOAT32,
@@ -121,14 +117,8 @@ var (
 			RangeExt: geocube.Range{Min: 0, Max: 1},
 			Exponent: 1,
 		},
-		{
-			DataFormat: geocube.DataFormat{
-				DType:  geocube.DTypeFLOAT32,
-				NoData: 0,
-				Range:  geocube.Range{Min: 0, Max: 1},
-			},
-			RangeExt: geocube.Range{Min: 0, Max: 1},
-			Exponent: 1,
-		},
+		imageDFormatFloat32NoData0,
+		imageDFormatFloat32NoData0,
+		imageDFormatFloat32NoData0,
 	}
 )