+++ docs/about/release-notes.md
@@ -14,3 +14,4 @@ consume a lot of local storage)
 - If a deletion task failed, the job must be in "DONEBUTUNTIDY" state
 - Dockerfile uses alpine3.16, same version as golane:alpine
 - GRPC message errors are limited to 3Kb
+- Dataset bands were not taken into account during warping